se agregan cupones y dashboard primera integracion

+++ src/main/java/com/ingeduardo/demostore/dto/OrderItemResponseDto.java
@@ -1,5 +1,6 @@
 package com.ingeduardo.demostore.dto;
 
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
@@ -7,5 +8,5 @@ public class OrderItemResponseDto {
     private String productId;
     private String productName;
     private Integer quantity;
-    private Double priceAtPurchase;
+    private BigDecimal priceAtPurchase;
 }
